+++ RELEASENOTES.md
@@ -2,6 +2,7 @@
 
 ### Version 1.0.6
 Release date 2021-04-19
+- **Add group member** from other country, issue #52.
 - **Help texts** added for *person*, *group*, *cargo types* and *meeting*.
 - **Page headings** now contains a *back*-button and a *help*-button if a help text is available.
 Layout of heading is also inproved.